removed stuipd changes

+++ app/Jobs/Server/Buoys/RemoveBuoy.php
@@ -46,6 +46,10 @@ class RemoveBuoy implements ShouldQueue
             $buoyService->removeBuoy($this->server, $this->buoy);
         });
 
+        if (! $this->wasSuccessful()) {
+            throw new ServerCommandFailed($this->getCommandErrors());
+        }
+
         $this->server->buoys()->detach($this->buoy);
     }
 }
